Change float to double

+++ MusicInstrument/Brass/BrassInstrument.java
@@ -4,7 +4,7 @@ import MusicInstrument.MusicInstrument;
 public class BrassInstrument extends MusicInstrument {
     private boolean keyed;
 
-    public BrassInstrument(String code, String produser, float price, float weight, boolean keyed) {
+    public BrassInstrument(String code, String produser, double price, double weight, boolean keyed) {
         super(code, produser, price, weight);
         this.keyed = keyed;
     }
