+++ MusicInstrument/Brass/Horn.java
@@ -2,7 +2,7 @@ package MusicInstrument.Brass;
 final public class Horn extends BrassInstrument {
     private char pitchKey;
 
-    public Horn(String code, String produser, float price, float weight, boolean keyed, char pitchKey) {
+    public Horn(String code, String produser, double price, double weight, boolean keyed, char pitchKey) {
         super(code, produser, price, weight, keyed);
         this.pitchKey = pitchKey;
     }
