+++ MusicInstrument/Percussion/PercussionInstrument.java
@@ -2,7 +2,7 @@ package MusicInstrument.Percussion;
 import MusicInstrument.MusicInstrument;
 
 public class PercussionInstrument extends MusicInstrument {
-    public PercussionInstrument(String code, String produser, float price, float weight) {
+    public PercussionInstrument(String code, String produser, double price, double weight) {
         super(code, produser, price, weight);
     }
 }
